add show endpoint; change money to integer

+++ db/migrate/20211025024417_create_invoices.rb
@@ -6,7 +6,7 @@ class CreateInvoices < ActiveRecord::Migration[6.1]
       t.string :number, null: false
       t.text :company, null: false
       t.text :billing, null: false
-      t.money :total_amount, null: false
+      t.integer :total_amount, null: false
       t.string :email_list, null: false
       t.timestamps
     end
